requirements.txt
Flask
requests
beautifulsoup4
readability-lxml
Flask-Caching
gunicorn